All models are implemented
With this commit, all models are implemented. The only thing left is to make the variables in the models more consistent. For example, the variable "title" in the Publication model is a CharField, while the variable "title" in the Book model is a TextField. This is because the title of a book can be longer than 255 characters, while the title of a publication is limited to 255 characters. This is not a problem, but it would be nice to make the variables more consistent.
There be a need to adjust the models in the future, but for now, the models are complete.