Data transer:
With current models you can transfer publications, persons, authorship, editorship and supervisorship.
Note that in this commit you have not yet accounted for verifying if the relation in the data is working correctly.
But you will properply check for this in your next commit.

The way your workflow is, is the following:
bash publications/implement_models.sh && python manage.py runscript publications.migration_script