This setup works
With this setup you can implement the data, using the models_that_looks_like_old_models.
In the datatransfer_script you modify the models a tiny bit, after the data has been inserted.

To transfer the data you simply run bash datatransfer_script.sh

The idea now is to seperate the problems into to. First problem is to transfer the data into
models that represents the old models as closely as possible.

The second problem is to refactor the models after the data is inserted. This is because it helps
me understand the models better, when i to refactor it. That will give me an advandtage, later on.