Big not easy to understand commit
This commit just made it work, the progress to make it work is lost.
This commit is being taking, with in mind, to focus on making mini version of this app,
that can be put in production asap.

It focuses on /pubs/publist/ page and /pubs/detail/ page.

The goal is to make it possible to view the details page and the list page.